Fix: Update and Rendering timing error.
There needs to be one initial "update" of simulations to go along
with the initial rendering of that scene. Without it, odd things
happen, especially if there are variables that don't get properly
initialized until the first run of update().

If this line is missing, than update() is skipped some number of
times before enough lag has accumilated to finally trigger the
update catch-up loop, and in the meantime render() is getting
called over and over with, potentially, nothing to process. The
whole simulation is perpetually behind real-time by one frame.